query-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.8
python-multipart==0.0.6 